+++ src/datalevin/spill.clj
@@ -9,16 +9,15 @@
    [taoensso.nippy :as nippy]
    [clojure.set :as set])
   (:import
-   [java.util Iterator List UUID NoSuchElementException Map]
+   [java.util Iterator List UUID NoSuchElementException Map Set Collection]
    [java.io DataInput DataOutput]
    [java.lang.management ManagementFactory]
    [javax.management NotificationEmitter NotificationListener Notification]
    [com.sun.management GarbageCollectionNotificationInfo]
-   [org.eclipse.collections.api.set.primitive MutableIntSet]
+   [org.eclipse.collections.impl.map.mutable UnifiedMap]
    [org.eclipse.collections.impl.list.mutable FastList]
-   [org.eclipse.collections.impl.map.mutable.primitive IntObjectHashMap]
    [clojure.lang ISeq IPersistentVector MapEntry Util Sequential
-    IPersistentMap MapEquivalence IFn]))
+    IPersistentMap MapEquivalence IObj IFn IPersistentSet]))
 
 (defonce memory-pressure (volatile! 0))
 
@@ -65,7 +64,8 @@
                           spill-dir
                           ^FastList memory
                           disk
-                          total]
+                          total
+                          ^:unsynchronized-mutable meta]
   ISpillable
 
   (memory-count ^long [_] (.size memory))
@@ -169,9 +169,7 @@
       (or (instance? Sequential other) (instance? List other))
       (if (not= (count this) (count other))
         false
-        (if (every? true? (map #(Util/equiv %1 %2) this other))
-          true
-          false))
+        (every? true? (map #(Util/equiv %1 %2) this other)))
 
       :else false))
 
@@ -204,7 +202,14 @@
             (throw (NoSuchElementException.)))))))
 
   IFn
+
   (invoke [this arg1] (.valAt this arg1))
+
+  IObj
+
+  (withMeta [this m] (set! meta m) this)
+
+  (meta [_] meta)
 
   Object
 
@@ -239,9 +244,7 @@
     (if (or (instance? List other) (instance? Sequential other))
       (if (not= (count this) (count other))
         false
-        (if (every? true? (map #(Util/equiv %1 %2) this other))
-          true
-          false))
+        (every? true? (map #(Util/equiv %1 %2) this other)))
       false)))
 
 (deftype RSVecSeq [^SpillableVector v
@@ -265,9 +268,7 @@
             (instance? Sequential other))
       (if (not= (count this) (count other))
         false
-        (if (every? true? (map #(Util/equiv %1 %2) this other))
-          true
-          false))
+        (every? true? (map #(Util/equiv %1 %2) this other)))
       false)))
 
 (defn new-spillable-vector
@@ -282,7 +283,8 @@
                                                  (volatile! nil)
                                                  (FastList.)
                                                  (volatile! nil)
-                                                 (volatile! 0))]
+                                                 (volatile! 0)
+                                                 nil)]
      (doseq [v vs] (.cons svec v))
      svec)))
 
@@ -301,20 +303,20 @@
     (dotimes [_ n] (.cons vs (nippy/thaw-from-in! in)))
     vs))
 
-(deftype SpillableIntObjMap [^long spill-threshold
-                             ^String spill-root
-                             spill-dir
-                             ^IntObjectHashMap memory
-                             disk]
+(deftype SpillableMap [^long spill-threshold
+                       ^String spill-root
+                       spill-dir
+                       ^UnifiedMap memory
+                       disk
+                       ^:unsynchronized-mutable meta]
   ISpillable
 
   (memory-count ^long [_] (.size memory))
 
-  (disk-count ^long [_]
-    (if @disk (l/entries @disk c/tmp-dbi) 0))
+  (disk-count ^long [_] (if @disk (l/entries @disk c/tmp-dbi) 0))
 
   (spill [this]
-    (let [dir (str spill-root "dtlv-spill-intobj-map-" (UUID/randomUUID))]
+    (let [dir (str spill-root "dtlv-spill-map-" (UUID/randomUUID))]
       (vreset! spill-dir dir)
       (vreset! disk (l/open-kv dir {:temp? true}))
       (l/open-dbi @disk c/tmp-dbi {:key-size Integer/BYTES}))
@@ -330,28 +332,26 @@
     (cond-> (.size memory) @disk (+ ^long (l/entries @disk c/tmp-dbi))))
 
   (containsKey [_ k]
-    (if (.containsKey memory (int k))
+    (if (.containsKey memory k)
       true
       (if @disk
-        (some? (l/get-value @disk c/tmp-dbi k :int))
+        (some? (l/get-value @disk c/tmp-dbi k))
         false)))
 
   (entryAt [_ k]
     (if-some [v (.get memory k)]
       (MapEntry. k v)
-      (when-some [v (l/get-value @disk c/tmp-dbi k :int)]
+      (when-some [v (l/get-value @disk c/tmp-dbi k)]
         (MapEntry. k v))))
 
   (valAt [_ k nf]
-    (if (integer? k)
-      (if-some [v (.get memory k)]
-        v
-        (if @disk
-          (if-some [v (l/get-value @disk c/tmp-dbi k :int)]
-            v
-            nf)
-          nf))
-      nf))
+    (if-some [v (.get memory k)]
+      v
+      (if @disk
+        (if-some [v (l/get-value @disk c/tmp-dbi k)]
+          v
+          nf)
+        nf)))
   (valAt [this k]
     (.valAt this k nil))
 
@@ -369,8 +369,8 @@
   MapEquivalence
 
   (keySet [_]
-    (set/union (set (.toArray (.keySet memory)))
-               (when @disk (set (l/key-range @disk c/tmp-dbi [:all] :int)))))
+    (set/union (set (.keySet memory))
+               (when @disk (set (l/key-range @disk c/tmp-dbi [:all])))))
 
   (equiv [this other]
     (cond
@@ -379,17 +379,14 @@
       (map? other)
       (if (not= (count this) (count other))
         false
-        (if (every? true?
-                    (map #(Util/equiv (get this %) (get other %))
-                         (.keySet this)))
-          true
-          false))
+        (every? true? (map #(Util/equiv (get this %) (get other %))
+                           (.keySet this))))
 
       :else false))
 
   (seq ^ISeq [this]
     (when (< 0 (count this))
-      (iterator-seq (.iterator this))))
+      (map #(MapEntry. % (.get this %)) (.keySet this))))
 
   Map
 
@@ -398,33 +395,41 @@
   (put [this k v]
     (if (< ^long @memory-pressure spill-threshold)
       (.put memory k v)
-      (do (when (nil? @disk) (.spill this))
-          (l/transact-kv @disk [[:put c/tmp-dbi k v :int]]))))
+      (when-not (= (.get memory k) v)
+        (when (nil? @disk) (.spill this))
+        (l/transact-kv @disk [[:put c/tmp-dbi k v]]))))
 
   (get [this k] (.valAt this k))
 
   (remove [_ k]
-    (if (.containsKey memory (int k))
+    (if (.containsKey memory k)
       (.remove memory k)
-      (when (and @disk (l/get-value @disk c/tmp-dbi k :int))
-        (l/transact-kv @disk [[:del c/tmp-dbi k :int]]))))
+      (when (and @disk (l/get-value @disk c/tmp-dbi k))
+        (l/transact-kv @disk [[:del c/tmp-dbi k]]))))
 
   (isEmpty [this] (= 0 (count this)))
 
   IFn
+
   (invoke [this arg1] (.valAt this arg1))
+
+  IObj
+
+  (withMeta [this m] (set! meta m) this)
+
+  (meta [_] meta)
 
   Iterable
 
   (iterator [this]
     (locking this
       (let [i        (volatile! 0)
-            kl       (.toList ^MutableIntSet (.keySet memory))
+            kl       (FastList. (.keySet memory))
             mn       (.size kl)
             db       @disk
             ^long dn (if db (l/entries db c/tmp-dbi) 0)
             dl       (when-not (zero? dn)
-                       (l/get-range db c/tmp-dbi [:all] :int))]
+                       (l/get-range db c/tmp-dbi [:all]))]
         (reify
           Iterator
           (hasNext [_]
@@ -452,17 +457,126 @@
       (l/close-kv @disk)
       (u/delete-files @spill-dir))))
 
-(defn new-spillable-intobj-map
-  ([] (new-spillable-intobj-map nil nil))
-  ([m] (new-spillable-intobj-map m nil))
+(defn new-spillable-map
+  ([] (new-spillable-map nil nil))
+  ([m] (new-spillable-map m nil))
   ([m {:keys [spill-threshold spill-root]
        :or   {spill-threshold c/+default-spill-threshold+
               spill-root      c/+default-spill-root+}}]
    (when (empty? @listeners) (memory-updater))
-   (let [smap (SpillableIntObjMap. spill-threshold
-                                   spill-root
-                                   (volatile! nil)
-                                   (IntObjectHashMap.)
-                                   (volatile! nil))]
+   (let [smap (SpillableMap. spill-threshold
+                             spill-root
+                             (volatile! nil)
+                             (UnifiedMap.)
+                             (volatile! nil)
+                             nil)]
      (doseq [[k v] m] (assoc smap k v))
      smap)))
+
+(nippy/extend-freeze
+  SpillableMap :spillable-map
+  [^SpillableMap x ^DataOutput out]
+  (nippy/freeze-to-out! out (into {} x)))
+
+(nippy/extend-thaw
+  :spillable-map
+  [^DataInput in]
+  (new-spillable-map (nippy/thaw-from-in! in)))
+
+(deftype SpillableSet [^SpillableMap impl
+                       ^:unsynchronized-mutable meta]
+  ISpillable
+
+  (memory-count ^long [_] (memory-count impl))
+
+  (disk-count ^long [_] (disk-count impl))
+
+  (spill [this] (spill impl) this)
+
+  IPersistentSet
+
+  (count [_] (count impl))
+
+  (contains[_ o] (.containsKey impl o))
+
+  (cons [this o] (.put impl o o) this)
+
+  (empty [this] (.empty impl) this)
+
+  (disjoin [this o] (.without impl o) this)
+
+  (equiv [this other]
+    (cond
+      (identical? this other) true
+
+      (set? other)
+      (if (not= (count this) (count other))
+        false
+        (every? true? (map #(.contains ^Set other %) (seq this))))
+
+      :else false))
+
+  (seq ^ISeq [_] (seq (.keySet impl)))
+
+  (get [_ o] (.valAt impl o))
+
+  Set
+
+  (size [_] (count impl))
+
+  (isEmpty [this] (= 0 (count this)))
+
+  IFn
+
+  (invoke [_ arg1] (.valAt impl arg1))
+
+  IObj
+
+  (withMeta [this m] (set! meta m) this)
+
+  (meta [_] meta)
+
+  Iterable
+
+  (iterator [_]
+    (let [iter (.iterator impl)]
+      (reify
+        Iterator
+        (hasNext [_] (.hasNext iter))
+        (next [_] (key (.next iter))))))
+
+  Collection
+
+  (toArray [this] (object-array (seq this)))
+
+  Object
+
+  (toString [_] (str (into #{} (keys impl))))
+
+  (finalize ^void [_] (.finalize impl)))
+
+(defn new-spillable-set
+  ([] (new-spillable-set nil nil))
+  ([s] (new-spillable-set s nil))
+  ([s {:keys [spill-threshold spill-root]
+       :or   {spill-threshold c/+default-spill-threshold+
+              spill-root      c/+default-spill-root+}}]
+   (when (empty? @listeners) (memory-updater))
+   (let [impl (SpillableMap. spill-threshold
+                             spill-root
+                             (volatile! nil)
+                             (UnifiedMap.)
+                             (volatile! nil)
+                             nil)]
+     (doseq [e s] (.put impl e e))
+     (SpillableSet. impl nil))))
+
+(nippy/extend-freeze
+  SpillableSet :spillable-set
+  [^SpillableSet x ^DataOutput out]
+  (nippy/freeze-to-out! out (into #{} x)))
+
+(nippy/extend-thaw
+  :spillable-set
+  [^DataInput in]
+  (new-spillable-set (nippy/thaw-from-in! in)))